Basic metadata for all data sets in maputils

This function returns a data frame with basic meta data for all data sets in maputils. This includes data object names, spatial domain, number of features, number of variables, ID column name, whether or not the SpatialPolygonsDataFrame has multilevel/overlapping features, and a short description.

mapdata()

Value

a data frame.

Examples

mapdata()

#> # A tibble: 10 x 7
#>    data    domain features variables id    multilevel description              
#>    <chr>   <chr>     <int>     <int> <chr> <lgl>      <chr>                    
#>  1 akcan   akcan         1         3 <NA>  F          PRISM domain mask        
#>  2 akcan2  akcan         6         3 NAME  F          PRISM domain state/provi~
#>  3 ecoreg1 ak            3         3 NAME  F          Alaska level 1 ecoregions
#>  4 ecoreg2 ak            9         3 NAME  F          Alaska level 2 ecoregions
#>  5 ecoreg3 ak           32         3 NAME  F          Alaska level 3 ecoregions
#>  6 aklcc   ak            5         4 NAME  F          Landscape Conservation C~
#>  7 lcc     akcan         5         4 NAME  F          Landscape Conservation C~
#>  8 cavm    ak            3         3 NAME  F          Alaska circumpolar artic~
#>  9 fmz     ak           14         5 NAME  F          Alaska Fire Service fire~
#> 10 tpa     akcan         8         5 NAME  F          Alaska/Canada terrestria~