Basic metadata for all apps in snapapps

This function returns a data frame with basic meta data for all Shiny apps in snapapps. This includes the following information:

snapps()

Value

a data frame.

Details

the app ID used for launching an app via snapp.

the app name.

a short description.

date originally published.

date most recently revised.

app complexity rating: beginner, intermediate, advanced or developer.

a status indicator of whether the app is complete or partial with respect to the canonical hosted app. For example, did the local app require available data sets to be reduced or removed the minimize its size in snapapps?

an indicator of whether the app relies on publicly accessible data stored on Amazon Web Services in order to minimize package size. When possible, AWS is used to ensure a packaged app can be offered as complete.

the canonical remote url.

the redirect url.

Examples

snapps()

#> # A tibble: 9 x 10
#>   id    name     descripti~ publi~ revis~ rati~ stat~ aws   url       redirect 
#>   <chr> <chr>    <chr>       <dbl>  <dbl> <fct> <chr> <lgl> <chr>     <chr>    
#> 1 rv    RV dist~ Distribut~   2017   2017 Begi~ comp~ F     https://~ http://s~
#> 2 twe   Tempera~ Coastal A~   2013   2017 Inte~ comp~ T     https://~ http://s~
#> 3 sic   Sea ice~ Alaska se~   2013   2017 Inte~ comp~ T     https://~ http://s~
#> 4 siw   Sea ice~ Coastal A~   2013   2017 Inte~ comp~ T     https://~ http://s~
#> 5 rv1   RV dist~ Distribut~   2013   2017 Begi~ comp~ F     https://~ http://s~
#> 6 rv2   RV dist~ Distribut~   2013   2017 Begi~ comp~ F     https://~ http://s~
#> 7 rv3   RV dist~ Distribut~   2013   2017 Begi~ comp~ F     https://~ http://s~
#> 8 rv4   RV dist~ Distribut~   2013   2017 Begi~ comp~ F     https://~ http://s~
#> 9 tring Tree ri~ Climate a~   2013   2017 Begi~ comp~ F     https://~ http://s~